+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-oci/compare/0.1.0...HEAD
 
+### Added
+
+-   Probe to count the number of instances in the compartment with the possibility of added filters.
+-   Associated documentation on how to use the probe.
+
 ## [0.1.0][]
 
 [0.1.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-oci/tree/0.1.0