+++ CHANGELOG.md
@@ -18,6 +18,11 @@
 -   Probe to count the number of instances in the compartment with the possibility of added filters.
 -   Associated documentation on how to use the probe.
 
+### Added
+
+-   Probe to count the number of instances in the compartment with the possibility of added filters.
+-   Associated documentation on how to use the probe.
+
 ## [0.1.0][]
 
 [0.1.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-oci/tree/0.1.0