+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-oci/compare/0.1.0...HEAD
 
+-   Stop random instance, allowing to do so in terms of the requested filters.
+
 ### Added
 
 -   Probe to count the number of instances in the compartment with the possibility of added filters.