+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-oci/compare/0.1.0...HEAD
 
--   Stop random instance, allowing to do so in terms of the requested filters.
+-   A script that provides the user with a list of probes and actions within the driver.
 
 ### Added
 
+-   Stop random instance, allowing to do so in terms of the requested filters.
 -   Probe to count the number of instances in the compartment with the possibility of added filters.
 -   Associated documentation on how to use the probe.